03-ReactBasics: 22. SearchField State
event.target.value to get the value from input box.

this.setState is asynchronous function, it does not happen immediately. As you see, the console.log after the this.setState is always one letter behind.
There is a callback function that we can pass as the second argument in this.setState.